Struct ncollide_pipeline::broad_phase::DBVTBroadPhase

pub struct DBVTBroadPhase<P: Point, BV, T> {
    // some fields omitted
}

Broad phase based on a Dynamic Bounding Volume Tree.

It uses two separate trees: one for static objects and which is never updated, and one for moving objects.

Methods

impl<P, BV, T> DBVTBroadPhase<P, BV, T> where P: Point, BV: 'static + BoundingVolume<P::Vect::Scalar> + Translation<P::Vect> + Clone

fn new(margin: P::Vect::Scalar, small_keys: bool) -> DBVTBroadPhase<P, BV, T>

Creates a new broad phase based on a Dynamic Bounding Volume Tree.

fn num_interferences(&self) -> usize

Number of interferences detected by this broad phase.