Module ncollide_pipeline::narrow_phase

Collision detection algorithms and structure for the Narrow Phase.

Structs

BallBall
Collision detector between two balls.

BasicCollisionDispatcher
Collision dispatcher for shapes defined by ncollide_entities.

CompositeShapeRepr
Collision detector between a concave shape and another shape.

ContactSignal
Signal for contact start/stop.

IncrementalContactManifoldGenerator
Contact manifold generator which keeps track of several contacts.

OneShotContactManifoldGenerator
Contact manifold generator producing a full manifold at the first update.

PlaneSupportMap
Collision detector between a plane and a shape implementing the SupportMap trait.

ReprCompositeShape
Collision detector between a shape and a concave shape.

SupportMapPlane
Collision detector between a plane and a shape implementing the SupportMap trait.

SupportMapSupportMap
Persistent collision detector between two shapes having a support mapping function.

Traits

CollisionDetector
Trait of the algorithms executed during the so-called Narrow Phase.

CollisionDispatcher

ContactSignalHandler
A signal handler for contact detection.

Type Definitions

CollisionAlgorithm