Struct ncollide_pipeline::narrow_phase::CompositeShapeRepr

pub struct CompositeShapeRepr<P: Point, M> {
    // some fields omitted
}

Collision detector between a concave shape and another shape.

Methods

impl<P: Point, M> CompositeShapeRepr<P, M>

fn new(prediction: P::Vect::Scalar) -> CompositeShapeRepr<P, M>

Creates a new collision detector between a concave shape and another shape.