Struct ncollide_pipeline::narrow_phase::OneShotContactManifoldGenerator

pub struct OneShotContactManifoldGenerator<N, P, V, CD> {
    // some fields omitted
}

Contact manifold generator producing a full manifold at the first update.

Whenever a new contact is detected (i.e. when the current manifold is empty) a full manifold is generated. Then, the manifold is incrementally updated by an IncrementalContactManifoldGenerator.

Methods

impl<N, P, V, M, CD> OneShotContactManifoldGenerator<N, P, V, CD> where CD: CollisionDetector<N, P, V, M>

fn new(prediction: N, cd: CD) -> OneShotContactManifoldGenerator<N, P, V, CD>

Creates a new one shot contact manifold generator.