Struct ncollide_pipeline::narrow_phase::ShapeCompositeShapeProximityDetector

pub struct ShapeCompositeShapeProximityDetector<P: Point, M> { /* fields omitted */ }

Proximity detector between a shape and a concave shape.

Methods

impl<P: Point, M> ShapeCompositeShapeProximityDetector<P, M>

fn new() -> ShapeCompositeShapeProximityDetector<P, M>

Creates a new collision detector between a shape and a concave shape.