Struct ncollide_pipeline::narrow_phase::SupportMapSupportMap

pub struct SupportMapSupportMap<N, P, V, M, S> {
    // some fields omitted
}

Persistent collision detector between two shapes having a support mapping function.

It is based on the GJK algorithm. This detector generates only one contact point. For a full manifold generation, see IncrementalContactManifoldGenerator.

Methods

impl<N, P, V, M, S> SupportMapSupportMap<N, P, V, M, S> where N: Scalar, P: Point<N, V>, V: Vect<N>, S: Simplex<N, AnnotatedPoint<P>>

fn new(prediction: N, simplex: S) -> SupportMapSupportMap<N, P, V, M, S>

Creates a new persistant collision detector between two shapes with support mapping functions.

It is initialized with a pre-created simplex.