Trait ncollide_pipeline::narrow_phase::CollisionDetector

pub trait CollisionDetector<N, P, V, M> {
    fn update(&mut self, &CollisionDispatcher<N, P, V, M>, &M, &Repr<N, P, V, M>, &M, &Repr<N, P, V, M>) -> bool;
    fn num_colls(&self) -> usize;
    fn colls(&self, &mut Vec<Contact<N, P, V>>);
}

Trait of the algorithms executed during the so-called Narrow Phase.

The goal of the narrow phase is to determine exactly if two objects collide. If there is collision, it must be able to compute the exact contact point(s), normal and penetration depth in order to give enough informations to the constraint solver.

Arguments

G1- the type of the first object involved on the collision detection.

G2- the type of the second object involved on the collision detection.

Required Methods

fn update(&mut self, &CollisionDispatcher<N, P, V, M>, &M, &Repr<N, P, V, M>, &M, &Repr<N, P, V, M>) -> bool

Runs the collision detection on two objects. It is assumed that the same collision detector (the same structure) is always used with the same pair of object.

fn num_colls(&self) -> usize

The number of collision detected during the last update.

fn colls(&self, &mut Vec<Contact<N, P, V>>)

Collects the collisions detected during the last update.