Trait ncollide_pipeline::narrow_phase::ProximityDetector

pub trait ProximityDetector<P: Point, M> {
    fn update(&mut self,
              dispatcher: &ProximityDispatcher<P, M>,
              ma: &M,
              a: &Shape<P, M>,
              mb: &M,
              b: &Shape<P, M>,
              margin: P::Real)
              -> bool;
    fn proximity(&self) -> Proximity;
}

Trait implemented by algorithms that determine if two objects are in close proximity.

Required Methods

fn update(&mut self,
dispatcher: &ProximityDispatcher<P, M>,
ma: &M,
a: &Shape<P, M>,
mb: &M,
b: &Shape<P, M>,
margin: P::Real)
-> bool

Runs the proximity detection on two objects. It is assumed that the same proximity detector (the same structure) is always used with the same pair of object.

fn proximity(&self) -> Proximity

The number of collision detected during the last update.