Struct ncollide_pipeline::world::CollisionObject

pub struct CollisionObject<P: Point, M, T> {
    pub uid: usize,
    pub position: M,
    pub shape: ShapeHandle<P, M>,
    pub collision_groups: CollisionGroups,
    pub query_type: GeometricQueryType<P::Real>,
    pub data: T,
    // some fields omitted
}

A stand-alone object that has a position and a shape.

Fields
uid: usize

The collsion object unique identifier.

position: M

The collision object position.

shape: ShapeHandle<P, M>

The collision object shape.

collision_groups: CollisionGroups

The collision groups of the collision object.

query_type: GeometricQueryType<P::Real>

The kind of queries this collision object is expected to .

data: T

The user-defined data associated to this object.

Methods

impl<P: Point, M, T> CollisionObject<P, M, T>

fn new(uid: usize,
position: M,
shape: ShapeHandle<P, M>,
groups: CollisionGroups,
query_type: GeometricQueryType<P::Real>,
data: T)
-> CollisionObject<P, M, T>

Creates a new collision object.