Struct ncollide_pipeline::world::CollisionObjectsDispatcher

pub struct CollisionObjectsDispatcher<N, P, V, M, T> {
    // some fields omitted
}

Collision detector dispatcher for collision objects.

Methods

impl<N, P, V, M, T> CollisionObjectsDispatcher<N, P, V, M, T>

fn new(shape_dispatcher: Box<CollisionDispatcher<N, P, V, M> + 'static>) -> CollisionObjectsDispatcher<N, P, V, M, T>

Creates a new CollisionObjectsDispatcher.

fn update(&mut self, objects: &UidRemap<CollisionObject<N, P, V, M, T>>, timestamp: usize)

Updates the contact pairs.

fn contact_pairs<F>(&self, objects: &UidRemap<CollisionObject<N, P, V, M, T>>, f: F) where F: FnMut(&T, &T, &CollisionAlgorithm<N, P, V, M>)

Iterates through all the contact pairs.

fn contacts<F>(&self, objects: &UidRemap<CollisionObject<N, P, V, M, T>>, f: F) where F: FnMut(&T, &T, &Contact<N, P, V>)

Calls a closures on each contact between two objects.

fn register_contact_signal_handler(&mut self, name: &str, handler: Box<ContactSignalHandler<T> + 'static>)

Registers a handler for contact start/stop events.

fn unregister_contact_signal_handler(&mut self, name: &str)

Unregisters a handler for contact start/stop events.

fn handle_proximity(&mut self, objects: &UidRemap<CollisionObject<N, P, V, M, T>>, fk1: &FastKey, fk2: &FastKey, started: bool)

Creates/removes the persistant collision detector associated to a given pair of objects.

fn is_proximity_allowed(objects: &UidRemap<CollisionObject<N, P, V, M, T>>, fk1: &FastKey, fk2: &FastKey) -> bool

Tests if two objects can be tested for mutual collision.