Struct ncollide_pipeline::world::CollisionWorld

pub struct CollisionWorld<P: Point, M, T> { /* fields omitted */ }

A world that handles collision objects.

Methods

impl<P, M, T> CollisionWorld<P, M, T> where P: Point,
P::Vect: Translate<P> + Cross,
P::Vect::CrossProductType: Vector<Scalar=P::Vect::Scalar> + Mul<P::Vect::Scalar, Output=P::Vect::CrossProductType>,
M: Isometry<P> + Translation<P::Vect> + Rotation<P::Vect::CrossProductType>

fn new(margin: P::Vect::Scalar, small_uids: bool) -> CollisionWorld<P, M, T>

Creates a new collision world.

fn deferred_add(&mut self,
uid: usize,
position: M,
shape: ShapeHandle<P, M>,
collision_groups: CollisionGroups,
query_type: GeometricQueryType<P::Vect::Scalar>,
data: T)

Adds a collision object to the world.

fn update(&mut self)

Updates the collision world.

This executes the whole collision detection pipeline: the broad phase first, then the narrow phase.

fn deferred_remove(&mut self, uid: usize)

Marks a collision object for removal from the world during the next update.

fn deferred_set_position(&mut self, uid: usize, pos: M)

Sets the position the collision object attached to the specified object will have during the next update.

fn register_broad_phase_pair_filter<F>(&mut self, name: &str, filter: F) where F: BroadPhasePairFilter<P, M, T> + 'static

Adds a filter that tells if a potential collision pair should be ignored or not.

The proximity filter returns false for a given pair of collision objects if they should be ignored by the narrow phase. Keep in mind that modifying the proximity filter will have a non-trivial overhead during the next update as it will force re-detection of all collision pairs.

fn unregister_broad_phase_pair_filter(&mut self, name: &str)

Removes the pair filter named name.

fn register_contact_handler<H>(&mut self, name: &str, handler: H) where H: ContactHandler<P, M, T> + 'static

Registers a handler for contact start/stop events.

fn unregister_contact_handler(&mut self, name: &str)

Unregisters a handler for contact start/stop events.

fn register_proximity_handler<H>(&mut self, name: &str, handler: H) where H: ProximityHandler<P, M, T> + 'static

Registers a handler for proximity status change events.

fn unregister_proximity_handler(&mut self, name: &str)

Unregisters a handler for proximity status change events.

fn perform_position_update(&mut self)

Executes the position updates.

fn perform_additions_removals_and_broad_phase(&mut self)

Actually adds or removes all the objects marked by .deferred_add(...) or .deferred_remove(...) and updates the broad phase.

fn perform_broad_phase(&mut self)

Executes the broad phase of the collision detection pipeline.

Note that this does not take in account the changes made to the collision updates after the last .perform_position_update() call.

fn perform_narrow_phase(&mut self)

Executes the narrow phase of the collision detection pipeline.

fn set_narrow_phase(&mut self,
narrow_phase: Box<NarrowPhase<P, M, T>>)
-> Box<NarrowPhase<P, M, T>>

Sets a new narrow phase and returns the previous one.

Keep in mind that modifying the narrow-pase will have a non-trivial overhead during the next update as it will force re-detection of all collision pairs and their associated contacts.

fn contact_pairs(&self) -> ContactPairs<P, M, T>

Iterates through all the contact pairs detected since the last update.

fn proximity_pairs(&self) -> ProximityPairs<P, M, T>

Iterates through all the proximity pairs detected since the last update.

fn contacts(&self) -> Contacts<P, M, T>

Iterates through every contact detected since the last update.

fn collision_objects(&self) -> Values<CollisionObject<P, M, T>>

Iterates through all collision objects.

fn collision_object(&self, uid: usize) -> Option<&CollisionObject<P, M, T>>

Returns a reference to the collision object identified by uid.

fn interferences_with_ray<'a>(&'a self,
ray: &'a Ray<P>,
groups: &'a CollisionGroups)
-> InterferencesWithRay<'a, P, M, T>

Computes the interferences between every rigid bodies on this world and a ray.

fn interferences_with_point<'a>(&'a self,
point: &'a P,
groups: &'a CollisionGroups)
-> InterferencesWithPoint<'a, P, M, T>

Computes the interferences between every rigid bodies of a given broad phase, and a point.

fn interferences_with_aabb<'a>(&'a self,
aabb: &'a AABB<P>,
groups: &'a CollisionGroups)
-> InterferencesWithAABB<'a, P, M, T>

Computes the interferences between every rigid bodies of a given broad phase, and a aabb.