Struct ncollide_pipeline::broad_phase::DBVTBroadPhase

pub struct DBVTBroadPhase<N, P, BV, T> {
    // some fields omitted
}

Broad phase based on a Dynamic Bounding Volume Tree.

It uses two separate trees: one for static objects and which is never updated, and one for moving objects.

Methods

impl<N, P, V, BV, T> DBVTBroadPhase<N, P, BV, T> where N: Scalar, P: Point<N, V>, V: Vect<N>, BV: 'static + BoundingVolume<N> + Translation<V> + Clone

fn new(margin: N, small_keys: bool) -> DBVTBroadPhase<N, P, BV, T>

Creates a new broad phase based on a Dynamic Bounding Volume Tree.

fn num_interferences(&self) -> usize

Number of interferences detected by this broad phase.