Struct ncollide_pipeline::narrow_phase::CompositeShapeRepr

pub struct CompositeShapeRepr<N, P, V, M> {
    // some fields omitted
}

Collision detector between a concave shape and another shape.

Methods

impl<N, P, V, M> CompositeShapeRepr<N, P, V, M>

fn new(prediction: N) -> CompositeShapeRepr<N, P, V, M>

Creates a new collision detector between a concave shape and another shape.