Struct ncollide_pipeline::world::CollisionGroups

pub struct CollisionGroups {
    // some fields omitted
}

30 groups of collision used to filter which object collide with which other one.

If two objects have at least one collision group in common, they can collide. Otherwise, they cannot.

Methods

impl CollisionGroups

fn new() -> CollisionGroups

Creates a new CollisionGroups that enables collisions with everything except self-collision.

fn new_with_groups(groups: &[usize]) -> CollisionGroups

Creates a new CollisionGroups that enables collisions with the given groups.

This does not enable self-collision.

fn new_without_groups(groups: &[usize]) -> CollisionGroups

Creates a new CollisionGroups that enables collisions with every group but the given ones.

This does not enable self-collision.

fn enable_collisions_with_all_groups(&mut self)

Enables collision with all groups.

This does not affect self-collision.

fn disable_collisions_with_all_groups(&mut self)

Disables collision with all groups.

This does not affect self-collision.

fn enable_collisions_with_group(&mut self, group_id: usize)

Enable collision detection with a specific group.

fn disable_collisions_with_group(&mut self, group_id: usize)

Disables collision detection with a specific group.

fn enable_self_collision(&mut self)

Enables self collision detection.

fn disable_self_collision(&mut self)

Disables self collision detection.

fn can_collide_with_group(&self, group_id: usize) -> bool

Tests whether collisions with a given group is enabled.

fn can_collide_with_groups(&self, other: &CollisionGroups) -> bool

Tests whether two collision groups have at least one group in common.

fn can_collide_with_self(&self) -> bool

Tests whether self-collision is enabled.