Struct ncollide_pipeline::world::CollisionObjectsDispatcher

pub struct CollisionObjectsDispatcher<P, M, T> {
    // some fields omitted
}

Collision detector dispatcher for collision objects.

Methods

impl<P: Point, M: 'static, T> CollisionObjectsDispatcher<P, M, T>

fn new(shape_dispatcher: Box<CollisionDispatcher<P, M> + 'static>) -> CollisionObjectsDispatcher<P, M, T>

Creates a new CollisionObjectsDispatcher.

fn update(&mut self, objects: &UidRemap<CollisionObject<P, M, T>>, timestamp: usize)

Updates the contact pairs.

fn contact_pairs<'a>(&'a self, objects: &'a UidRemap<CollisionObject<P, M, T>>) -> ContactPairs<'a, P, M, T>

Iterates through all the contact pairs.

fn contacts<'a>(&'a self, objects: &'a UidRemap<CollisionObject<P, M, T>>) -> Contacts<'a, P, M, T>

Iterates through all the contacts detected since the last update.

fn register_contact_signal_handler(&mut self, name: &str, handler: Box<ContactSignalHandler<T> + 'static>)

Registers a handler for contact start/stop events.

fn unregister_contact_signal_handler(&mut self, name: &str)

Unregisters a handler for contact start/stop events.

fn handle_proximity(&mut self, objects: &UidRemap<CollisionObject<P, M, T>>, fk1: &FastKey, fk2: &FastKey, started: bool)

Creates/removes the persistant collision detector associated to a given pair of objects.

fn is_proximity_allowed(objects: &UidRemap<CollisionObject<P, M, T>>, fk1: &FastKey, fk2: &FastKey) -> bool

Tests if two objects can be tested for mutual collision.