Struct ncollide_pipeline::world::CollisionWorld

pub struct CollisionWorld<N, P, V, M, T> {
    // some fields omitted
}

A world that handles collision objects.

Methods

impl<N, P, V, AV, M, T> CollisionWorld<N, P, V, M, T> where N: Scalar, P: Point<N, V>, V: Vect<N> + Translate<P> + Cross<Output=AV>, AV: Vect<N>, M: Isometry<N, P, V> + Rotation<AV>

fn new(margin: N, prediction: N, small_uids: bool) -> CollisionWorld<N, P, V, M, T>

Creates a new collision world.

fn add(&mut self, uid: usize, position: M, shape: Arc<Box<Repr<N, P, V, M>>>, collision_groups: CollisionGroups, data: T)

Adds a collision object to the world.

fn remove(&mut self, uid: usize)

Remove a collision object from the world.

fn update(&mut self)

Updates the collision world.

This executes the whole collision detection pipeline: the broad phase first, then the narrow phase.

fn defered_set_position(&mut self, uid: usize, pos: M)

Sets the position the collision object attached to the specified object will have during the next update.

fn register_contact_signal_handler<H>(&mut self, name: &str, handler: H) where H: ContactSignalHandler<T> + 'static

Registers a handler for contact start/stop events.

fn unregister_contact_signal_handler(&mut self, name: &str)

Unregisters a handler for contact start/stop events.

fn perform_position_update(&mut self)

Executes the position updates.

fn perform_broad_phase(&mut self)

Executes the broad phase of the collision detection pipeline.

Not that this does not take in account the changes made to the collision updates after the last .perform_position_update() call.

fn perform_narrow_phase(&mut self)

Executes the narrow phase of the collision detection pipeline.

fn contact_pairs<F>(&self, f: F) where F: FnMut(&T, &T, &CollisionAlgorithm<N, P, V, M>)

Iterats through all the contact pairs.

fn contacts<F>(&self, f: F) where F: FnMut(&T, &T, &Contact<N, P, V>)

Collects every contact detected since the last update.

fn interferences_with_ray<'a, F>(&'a mut self, ray: &Ray<P, V>, f: F) where F: FnMut(&T, RayIntersection<N, V>)

Computes the interferences between every rigid bodies of a given broad phase, and a ray.

fn interferences_with_point<F>(&self, point: &P, f: F) where F: FnMut(&T)

Computes the interferences between every rigid bodies of a given broad phase, and a point.

fn interferences_with_aabb<F>(&self, aabb: &AABB<P>, f: F) where F: FnMut(&T)

Computes the interferences between every rigid bodies of a given broad phase, and a aabb.