Struct ncollide_pipeline::world::CollisionWorld

pub struct CollisionWorld<P, M, T> {
    // some fields omitted
}

A world that handles collision objects.

Methods

impl<P, M, T> CollisionWorld<P, M, T> where P: Point, P::Vect: Translate<P> + Cross, P::Vect::CrossProductType: Vect<Scalar=P::Vect::Scalar> + Mul<P::Vect::Scalar, Output=P::Vect::CrossProductType>, M: Isometry<P, P::Vect> + Translation<P::Vect> + Rotation<P::Vect::CrossProductType>

fn new(margin: P::Vect::Scalar, prediction: P::Vect::Scalar, small_uids: bool) -> CollisionWorld<P, M, T>

Creates a new collision world.

fn add(&mut self, uid: usize, position: M, shape: Arc<Box<Repr<P, M>>>, collision_groups: CollisionGroups, data: T)

Adds a collision object to the world.

fn update(&mut self)

Updates the collision world.

This executes the whole collision detection pipeline: the broad phase first, then the narrow phase.

fn deferred_remove(&mut self, uid: usize)

Marks a collision object for removal from the world during the next update.

fn deferred_set_position(&mut self, uid: usize, pos: M)

Sets the position the collision object attached to the specified object will have during the next update.

fn register_contact_signal_handler<H>(&mut self, name: &str, handler: H) where H: ContactSignalHandler<T> + 'static

Registers a handler for contact start/stop events.

fn unregister_contact_signal_handler(&mut self, name: &str)

Unregisters a handler for contact start/stop events.

fn perform_position_update(&mut self)

Executes the position updates.

fn perform_broad_phase(&mut self)

Executes the broad phase of the collision detection pipeline.

Not that this does not take in account the changes made to the collision updates after the last .perform_position_update() call.

fn perform_removals_and_broad_phase(&mut self)

Actually removes all the objects marked by .deferred_remove(...) and updates the broad phase.

fn perform_narrow_phase(&mut self)

Executes the narrow phase of the collision detection pipeline.

fn contact_pairs(&self) -> ContactPairs<P, M, T>

Iterates through all the contact pairs detected since the last update.

fn contacts(&self) -> Contacts<P, M, T>

Iterates through every contact detected since the last update.

fn interferences_with_ray<'a>(&'a self, ray: &'a Ray<P>, groups: &'a CollisionGroups) -> InterferencesWithRay<'a, P, M, T>

Computes the interferences between every rigid bodies on this world and a ray.

fn interferences_with_point<'a>(&'a self, point: &'a P, groups: &'a CollisionGroups) -> InterferencesWithPoint<'a, P, M, T>

Computes the interferences between every rigid bodies of a given broad phase, and a point.

fn interferences_with_aabb<'a>(&'a self, aabb: &'a AABB<P>, groups: &'a CollisionGroups) -> InterferencesWithAABB<'a, P, M, T>

Computes the interferences between every rigid bodies of a given broad phase, and a aabb.